Move files to near-network-private-actix
