+++ chain/network/src/lib.rs
@@ -12,6 +12,10 @@ pub use near_network_primitives::types::PeerInfo;
 mod network_protocol;
 mod peer;
 mod peer_manager;
+#[cfg(feature = "test_features")]
+pub mod private_actix;
+#[cfg(not(feature = "test_features"))]
+pub(crate) mod private_actix;
 pub mod routing;
 pub(crate) mod stats;
 pub mod test_utils;
