+++ chain/network/src/routing/mod.rs
@@ -10,13 +10,13 @@ mod route_back_cache;
 pub(crate) mod routing;
 pub(crate) mod routing_table_actor;
 
+pub use crate::network_protocol::{Edge, EdgeState, PartialEdgeInfo, SimpleEdge};
+#[cfg(feature = "test_features")]
+pub use crate::private_actix::GetRoutingTableResult;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub use crate::routing::ibf_peer_set::SlotMapId;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub use crate::routing::ibf_set::IbfSet;
-pub use crate::routing::network_protocol::{EdgeState, SimpleEdge};
-#[cfg(feature = "test_features")]
-pub use crate::routing::routing::GetRoutingTableResult;
 pub use crate::routing::routing::{
     Graph, RoutingTableView, DELETE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME,
 };
